web_app/backend/requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing & Data
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0

# Machine Learning & AI
tensorflow>=2.14.0
scikit-learn>=1.3.0
albumentations>=1.3.0

# Presentation Generation
python-pptx>=0.6.21
